% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankWorkSheet.fn.R
\name{BlankWorkSheet}
\alias{BlankWorkSheet}
\title{Blank Worksheet}
\usage{
BlankWorkSheet(wb, sheet, skipEmptyRows = TRUE)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{sheet}{name or index of worksheet to create or clear.}
}
\value{
updates the openxlsx workbook object.
}
\description{
Creates either a blank worksheet if one does not exists or clears out
the worksheet if it already is present in the Openxlsx workbook.
}
\examples{

\dontrun{
# Add a new worksheet to a new work book
wb <- createWorkbook()
BlankWorksheet(wb, "TEST")

# Clear out an existing worksheet
wb <- read.xlsx("path/to/workbook")
BlankWorksheet(wb, "ExistingSheetName")
}

}
