% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAdults.fn.R
\name{GetAdults}
\alias{GetAdults}
\title{Get Adults Results from Simulation}
\usage{
GetAdults(sim.result, type = "performance", scenario = NULL)
}
\arguments{
\item{sim.result}{}

\item{type}{}

\item{scenario}{}
}
\description{
Get Adults Results from Simulation
}
