% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSettings.fn.R
\name{GetSettings}
\alias{GetSettings}
\title{Get Simulation Settings}
\usage{
GetSettings(type = "default v2")
}
\arguments{
\item{type}{character string indicating the settings scenario to retrieve}
}
\value{
list object containing required simulator parameter settings.
}
\description{
Get Simulation Settings
}
\examples{
setting.default <- GetSettings("default")
str(setting.default)

}
