% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpawnerRecruits.fn.R
\name{GetSpawnerRecruits}
\alias{GetSpawnerRecruits}
\title{Get Simulation Spawner Recruits}
\usage{
GetSpawnerRecruits(sim.result, type = "summary", scenario = NULL)
}
\arguments{
\item{sim.result}{}

\item{type}{}
}
\description{
Get Simulation Spawner Recruits
}
