% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpawners.fn.R
\name{GetSpawners}
\alias{GetSpawners}
\title{Get Simulation Spawners}
\usage{
GetSpawners(sim.result, type = "summary", min.yr = 2006, scenario = NULL)
}
\arguments{
\item{sim.result}{}

\item{type}{}
}
\description{
Get Simulation Spawners
}
