% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCCRiskCat.fn.R
\name{IPCCRiskCat}
\alias{IPCCRiskCat}
\title{IPCC Risk Category}
\usage{
IPCCRiskCat(p)
}
\arguments{
\item{p}{numeric value containing the outcome probability.}
}
\value{
A character vector containing the IPCC Risk Category
}
\description{
IPCC Risk Category
}
\details{

}
\examples{

}
\references{
Mastrandrea, M. D., C. B. Field, T. F. Stocker, O. Edenhofer, K. L. Ebi, D. J. Frame, H. Held, E. Kriegler, K. J. Mach, P. R. Matschoss, G.-K. Plattner, G. W. Yohe, and F. W. Zwiers. 2010. Guidance note for lead authors of the IPCC fifth assessment report on consistent treatment of uncertainties. Intergovernmental Panel on Climate Change (IPCC).
}
