% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAdultMetrics.fn.R
\name{PlotAdultMetrics}
\alias{PlotAdultMetrics}
\title{Plot Adult Performance Metrics}
\usage{
PlotAdultMetrics(runs, bins = 45, return = FALSE)
}
\arguments{
\item{runs}{list object containing individual simulation runs, element names are used for panel labels.}

\item{bins}{numeric indicating the number of bins to display in the histograms}

\item{return}{logical indicating}
}
\value{
displays a plot if  `return = FALSE` (default), otherwise return a grob without drawing.
}
\description{
Plot Adult Performance Metrics
}
\examples{
# Default  simulation settings
runs <- list()
runs[['Default']] <- PopSim(GetSettings('default'), n.sim=100)

# Improve Early Ocean, Ocean Harvest
settings <- GetSettings('default')
settings$hr.ocean <- settings$hr.ocean /2
settings$S.osar = 0.3
runs[['Improved Early Ocean Survival + Reduced Ocean Harvest']] <- PopSim(settings, n.sim=100)

# Supplementation
settings <- GetSettings('default')
settings$supplementation <-  1.75 * 10^6
settings$suppl.fitness <- 1
runs[['Supplementation (1.75 M)']] <- PopSim(settings, n.sim=100)

PlotMetrics(runs)
}
