% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetricECDF.fn.R
\name{PlotMetricECDF}
\alias{PlotMetricECDF}
\title{Plot Simulated Spawners}
\usage{
PlotMetricECDF(runs, type = "MPS", panel.widths = c(1, 1.35),
  return.plot = FALSE)
}
\arguments{
\item{runs}{named list object containing simulation runs to plot.}

\item{return}{}
}
\description{
Plot Simulated Spawners
}
