% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpawnerRecruits.fn.R
\name{PlotSpawnerRecruits}
\alias{PlotSpawnerRecruits}
\title{Plot Simulated Spawner Recruits}
\usage{
PlotSpawnerRecruits(runs, return = FALSE)
}
\arguments{
\item{runs}{named list object containing simulation runs to plot.}

\item{return}{}
}
\description{
Plot Simulated Spawner Recruits
}
