% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpawners.fn.R
\name{PlotSpawners}
\alias{PlotSpawners}
\title{Plot Simulated Spawners}
\usage{
PlotSpawners(runs, return = FALSE)
}
\arguments{
\item{runs}{named list object containing simulation runs to plot.}

\item{return}{}
}
\description{
Plot Simulated Spawners
}
