% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSim.fn.R
\name{PopSim}
\alias{PopSim}
\title{Chinook Population Simulator}
\usage{
PopSim(sim.settings = NULL, years = 2020:2060, n.sim = 10000,
  verbose = FALSE)
}
\arguments{
\item{sim.settings}{list object containing simulation parameter values, see details for more information.}

\item{n.sim}{numeric containing number of simulation runs.}

\item{verbose}{logical indicating whether status messages should be displayed.}
}
\value{
a list object with simulation results.
}
\description{
Chinook Population Simulator
}
\examples{
sim.result <- PopSim(sim.settings = GetSettings("default"), n.sim=5) 
str(sim.result)

}
